requirements.txt
python-dotenv
azure-openai
httpx
pillow
